upgrade to latest staticcheck
In latest staticcheck version, `staticcheck.Analyzers` and
`stylecheck.Analyzers` have been turned from maps to slices.  We need to
adjust the code accordingly so that we no longer concating 2 maps by
their keys.

After chatting with @dominikh, it seems that analysis.Analyzer does have
a 'Name' field exported that we can use for package name.

Remove unused `Analyzer` field in the anonymous data struct that we are
feeding to templates.

Adjusted doc in README accordingly.